gudangObatv2/requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1605523041333/work
blinker==1.4
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1606601387305/work
chardet @ file:///D:/bld/chardet_1607673982459/work
cryptography @ file:///D:/bld/cryptography_1607606473081/work
defusedxml==0.6.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1606808783925/work
django-allauth==0.41.0
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1593129968237/work
django-filter @ file:///home/conda/feedstock_root/build_artifacts/django-filter_1601214336459/work
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1601312268278/work
gunicorn==20.0.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///D:/bld/importlib-metadata_1607976164683/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1603668500820/work
oauthlib==3.0.1
psycopg2==2.8.6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
PyJWT==1.7.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
PySocks @ file:///D:/bld/pysocks_1602327203253/work
python3-openid @ file:///D:/bld/python3-openid_1602327014813/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1608904108784/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1602142927465/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1605225765842/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1602260267949/work
wincertstore==0.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
